requirements.txt
Flask==0.12.2
gunicorn==19.6.0
line-bot-sdk==1.3.0
Pillow==5.1.0
numpy==1.14.2
